+++ app/Http/Controllers/UsersController.php
@@ -5,22 +5,12 @@ use Illuminate\Http\Request;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
 use JobApis\JobsToMail\Http\Requests\CreateUser;
-use JobApis\JobsToMail\Jobs\ConfirmUser;
 use JobApis\JobsToMail\Jobs\CreateUserAndSearch;
 use JobApis\JobsToMail\Jobs\DeleteUser;
-use JobApis\JobsToMail\Jobs\GetUserSearches;
 
 class UsersController extends BaseController
 {
     use DispatchesJobs, ValidatesRequests;
-
-    /**
-     * Home page and signup form
-     */
-    public function index()
-    {
-        return view('users.index');
-    }
 
     /**
      * Create new User.
@@ -37,39 +27,13 @@ class UsersController extends BaseController
     }
 
     /**
-     * Confirm user account
+     * Delete a user account (unsubscribe from all searches)
      */
-    public function confirm(Request $request, $token)
-    {
-        $message = $this->dispatchNow(new ConfirmUser($token));
-
-        $request->session()->flash($message->type, $message->message);
-
-        return redirect('/');
-    }
-
-    /**
-     * Unsubscribe user account
-     */
-    public function unsubscribe(Request $request, $userId)
+    public function delete(Request $request, $userId)
     {
         $message = $this->dispatchNow(new DeleteUser($userId));
 
         $request->session()->flash($message->type, $message->message);
-
-        return redirect('/');
-    }
-
-    /**
-     * View searches for this user
-     */
-    public function searches(Request $request, $userId)
-    {
-        $results = $this->dispatchNow(new GetUserSearches($userId));
-
-        if (!$results->isEmpty()) {
-            return view('searches.index', ['searches' => $results]);
-        }
 
         return redirect('/');
     }