+++ app/Http/Requests/LoginUser.php
@@ -1,0 +1,41 @@
+<?php namespace JobApis\JobsToMail\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class LoginUser extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get custom messages for validator errors.
+     *
+     * @return array
+     */
+    public function messages()
+    {
+        return [
+            'exists' =>'This email address is not registered. 
+                Please try another email or create a new account.'
+        ];
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'email' => 'required|email|exists:users,email,deleted_at,NULL',
+        ];
+    }
+}