+++ resources/views/layouts/footer.blade.php
@@ -1,6 +1,4 @@
 <footer>
     &copy; {{ date('Y') }}
-    <a href="http://www.jobapis.com/" target="_blank">JobApis</a>
-    | <a href="https://twitter.com/jobapis" target="_blank">Twitter</a>
-    | <a href="https://github.com/jobapis" target="_blank">Github</a>
+    Created by <a href="http://www.jobapis.com/" target="_blank">JobApis</a>. Follow us on <a href="https://twitter.com/jobapis" target="_blank">Twitter</a> and <a href="https://github.com/jobapis" target="_blank">Github</a>.
 </footer>