+++ resources/views/users/index.blade.php
@@ -3,15 +3,21 @@
 @section('title', config('app.description'))
 
 @section('content')
+
+@include('layouts.sitetitle')
+
 <div class="row signup">
     <div class="col-lg-6 offset-lg-3 col-sm-8 offset-sm-2">
-        <h1 class="page-header">{{ config('app.name') }}</h1>
-        <p class="lead">{{ config('app.description') }}</p>
         <form method="POST" action="/users">
             {{ csrf_field() }}
             <div class="form-group">
                 <label for="email">Email</label>
-                <input type="email" name="email" class="form-control" placeholder="youremail@example.com" required/>
+                @if (session('user'))
+                    <input type="email" name="email" class="form-control" value="{{ session('user.email') }}" readonly />
+                    <small>You are currently logged in. <a href="/logout">Logout</a> to create a search as a new user.</small>
+                @else
+                    <input type="email" name="email" class="form-control" placeholder="youremail@example.com" required/>
+                @endif
             </div>
             <div class="form-group">
                 <label for="keyword">Search Term</label>
@@ -32,7 +38,11 @@
                 </label>
             </div>
             <div class="form-group">
-                <input type="submit" value="Sign Up" class="form-control btn btn-success btn-lg"/>
+                @if (session('user'))
+                    <input type="submit" value="Add Search" class="form-control btn btn-success btn-lg"/>
+                @else
+                    <input type="submit" value="Sign Up" class="form-control btn btn-success btn-lg"/>
+                @endif
                 <small><a href="/terms">Terms/Privacy Policy</a></small>
             </div>
         </form>