+++ changelog.md
@@ -4,11 +4,9 @@ All Notable changes to `jobs-to-mail` will be documented in this file.
 ## 1.0.0 - WIP
 
 ### Added
-- New database table/model for Searches.
+- User Logins via email/token.
 - Ability for Users to confirm emails once and create multiple Searches.
-- `jobs:email` command now uses Search model instead of User model for queries.
 - Storing daily email data in `notifications` DB table.
-- `DatabaseTruncater` class truncates `notifications`.
 - New links to email footer to manage job search subscriptions:
   - Unsubscribe from individual search
   - Unsubscribe from all searches
@@ -18,6 +16,8 @@ All Notable changes to `jobs-to-mail` will be documented in this file.
   - `recruiters` database table, seeder, model.
   - Boolean value for filtering recruiters from results of a search.
   - New filter in SearchAndNotifyUser job to remove recruiter listings if preferred.
+- New database table/model for Searches.
+- `jobs:email` command now uses Search model instead of User model for queries.
 - Moved job/collection-related filters into their own folder.
 - Added JS dependencies: Jquery, Bootstrap, Tether
 - Support for [Jobinventory](https://github.com/jobapis/jobs-jobinventory) job board.